--- a/Package.swift
+++ b/Package.swift
@@ -21,6 +21,7 @@ var sources = [
     "ggml/src/ggml-threading.cpp",
     "ggml/src/ggml-quants.c",
     "ggml/src/ggml-fp8.cpp",
+    "ggml/src/ggml-cpu/ggml-cpu-fp8.cpp",
 ]
 
 var resources: [Resource] = []
@@ -89,5 +90,5 @@ let package = Package(
             linkerSettings: linkerSettings
         )
     ],
-    cxxLanguageStandard: .cxx11
+    cxxLanguageStandard: .cxx17
 )