After make many test with FP8 this is the first part for add FP8 on llama.cpp.

I made some choose:
I implement 4 FP8 support:
- E5M2 & E4M3: for use with FP8 distributed model
it is not for create quantized model from FP16/BF16 model, but for FP8 distributed model like [Llama-3.1-405B-FP8](https://huggingface.co/meta-llama/Llama-3.1-405B-FP8) so I only add element for run.

- E4M3_Q & E3M4_Q: for gguf quantized model.
With this on we can quantize model I made test with Mistral-Nemo / Mistral-7B and Llama-3.1-8B .
the weight have simple quantization process like with Q8_0, but with size bloc of 256 (Q8_0 use bloc of 32)

For now all is in place for CPU backend as reference for any arch, so only use pure C++. I use openmp-simd for a little speed-up. 

I add 2 files gguf-fp8.h gguf-fp8.cpp to a lower change on existing sources, and simple merge.
For now I update Makefile build (need to figure what to do with CMake)

I prefer to use C++ for template on FP8 so I can have the 4 formats with little code size. 

So for now any comment is welcome.
Next step is to make a true merge request, before add faster compute. 

- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [ ] Low
  - [x] Medium
  - [ ] High

For those curious I have a high speed version I use for simple test [here](https://github.com/Djip007/llamafile/tree/feature/fp8) for zen4. And we have more discussion [here](https://github.com/Mozilla-Ocho/llamafile/discussions/549)