@Djip007 2024-10-28
> What are the advantages of the `FP8_Q` types, for example compared to the existing `Q8_0` type?

Some possible advantages:
- Q8_0 use bloc of 32 FP8_Q can use bloc of 256, so get a bpw of 8.125 vs 8.5.
- with CPU like zen4 we can use BF16 dot2 for better speed as we have with Q8_0 on CPU. (?)
- On supported hardware we can use FP8 tensor core with only need to apply the scale after 256 weight vs 32 for Q8_0, 
- ...

some advantage over FP8 is:
- with this FP8_Q to use native FP8 tensor core, we need to get the max for only 256 bloc elements for convert B to FP8, when we need the max for full row with "simple" FP8
- On my experiment the FP8_Q can be use with more accuracy that with FP8, so can be use with more model

Note: In fact I have some good accuracy with bloc of 512 or 1024, but I prefer to keep the same bloc size that is use for other Qn_K quant.