@Djip007 2024-10-29
OK have hard time with macOS compiler... 
I think I finally remove the need for c++17 (and make a little clean-up) :crossed_fingers: 

thanks @slaren @ggerganov