@ggerganov 2024-10-29
Got it. So the native `FP8` types are useful for models like https://huggingface.co/meta-llama/Llama-3.1-405B-FP8 and https://huggingface.co/neuralmagic/Mistral-Nemo-Instruct-2407-FP8 which are already trained in this format. While `FP8_Q` types can be used to quantize `BF16` models into data format that can utilize the NVIDIA FP8 tensor cores.