@Djip007 2024-10-29
Yes, that's how I see it.

- For FP8_Q I update [quantize.cpp](https://github.com/ggerganov/llama.cpp/pull/10055/files#diff-6745585c496560d324d1f0d6d77beebcb6dd9c3354bef41ab262535a87a376a7) so we can convert FP16/BF16.gguf to FP8_Q and use it.

- For FP8 I'm not sure what to do (I also didn't really look..) .
They (for [Llama-3.1-405B-FP8](https://huggingface.co/meta-llama/Llama-3.1-405B-FP8)) add a `weight_scale` with weight/bias on OP. It have to be apply like the bias after compute (bias is a add, scale is a mul).