@sorasoras 2024-11-24
> > > @0cc4m hi. is it possible to save more vram by using `VK_KHR_8bit_storage` after llama.cpp got fp8 support?
> > 
> > We're already using that extension to support the other quants, but it's only for 8-bit integers, not floats. FP8 is not currently supported by Vulkan at all as far as I can see.
> 
> is there any plan to support fp8 in vulkan backend in the future?

I don't see how. the problem is vulkan itself don't support fp8