@Djip007 2024-11-29
rebase on last master, with little change on fp32->fp8 rounding (add round to nearest even)