@Djip007 2024-11-29
> I am also not sure that we should add the OpenMP SIMD dependency. I don't really understand how it is supposed to work. Does the compiler really produce better code when it is enabled?

```
# without OpenMP SIMD
llama_perf_context_print: prompt eval time =   38115,07 ms /    33 tokens ( 1155,00 ms per token,     0,87 tokens per second)
llama_perf_context_print:        eval time =   19921,22 ms /    15 runs   ( 1328,08 ms per token,     0,75 tokens per second)

# with OpenMP SIMD
llama_perf_context_print: prompt eval time =   12686,11 ms /    33 tokens (  384,43 ms per token,     2,60 tokens per second)
llama_perf_context_print:        eval time =    6604,80 ms /    15 runs   (  440,32 ms per token,     2,27 tokens per second)
```
So yes it help. But it is true we can have much better than that with more optimised intrinsic code. So as you like.
I will add hight speed CPU matmul OP after this PR is merge.