@slaren 2024-11-29
> So yes it help. But it is true we can have much better than that with more optimised intrinsic code. So as you like.
> I will add hight speed CPU matmul OP after this PR is merge.

We can keep it until it is replaced with intrinsics then. It should still be limited to only the CPU backend, though.