@Djip007 2024-11-29
> We can keep it until it is replaced with intrinsics then. It should still be limited to only the CPU backend, though.

OK, Will remove this openmp-simd.

> Only the conversion code should be in ggml base, the vec dot code needs to be built in the CPU backend since that's the only code optimized for the CPU arch.

Good catch, i did it before the move of cpu in backend.
I'll have a look at it. an try to split the ggml-fp8 with ggml-cpu-fp8