@Djip007 2024-11-30
OK rebase with later master and now it is possible made use of c++17.

I refactor the fp8 so dot is now in cpu-backend. There is a few copie for now but may be remove with optimized code. I add a optimised dot for EnMn_Q and some arch (can't test on all hardware, work on x86 (AVX2/AVX512)

```
# without OpenMP SIMD
llama_perf_context_print: prompt eval time =   38115,07 ms /    33 tokens ( 1155,00 ms per token,     0,87 tokens per second)
llama_perf_context_print:        eval time =   19921,22 ms /    15 runs   ( 1328,08 ms per token,     0,75 tokens per second)

# with OpenMP SIMD
llama_perf_context_print: prompt eval time =   12686,11 ms /    33 tokens (  384,43 ms per token,     2,60 tokens per second)
llama_perf_context_print:        eval time =    6604,80 ms /    15 runs   (  440,32 ms per token,     2,27 tokens per second)

# with AVX512:
llama_perf_context_print: prompt eval time =    4507,33 ms /    33 tokens (  136,59 ms per token,     7,32 tokens per second)
llama_perf_context_print:        eval time =    3207,50 ms /    15 runs   (  213,83 ms per token,     4,68 tokens per second)

# Q8_0:
llama_perf_context_print: prompt eval time =     887,87 ms /    33 tokens (   26,91 ms per token,    37,17 tokens per second)
llama_perf_context_print:        eval time =    3244,96 ms /    15 runs   (  216,33 ms per token,     4,62 tokens per second)
```
at least the tg is in line with the Q8 on my CPU.