@slaren 2024-12-01
> * E5M2 & E4M3: for use with FP8 distributed model
>       it is not for create quantized model from FP16/BF16 model, but for FP8 distributed model like [Llama-3.1-405B-FP8](https://huggingface.co/meta-llama/Llama-3.1-405B-FP8) so I only add element for run.

As far as I can tell, this conversion is not implemented, so it is not actually possible to use FP8 models.