@Djip007 2024-12-01
> I have taken a look at the available FP8 models, and they use a single F32 scale per weight (edit: it's per row actually). This is something that is not supported in our quantization schemes and would require significant changes to implement.

For this case we don't need to add it in quantization schemes, but can add it in model load ("just" add mul op with it after the matmul, like with bias.)?

> So as it is, this implementation cannot be used with FP8 models, and the FP8 Q types perform worse than Q8_0. And I don't think that having a different group size than Q8_0 is not a good reason to add them either. I am not sure that there are any practical applications for this code as it is at the moment.

FP8_Q is perfectly usable and with good result.
what I get with Mistral_nemo:

|   BF16   |    PPL    |    KLD   | top P   |
| -------: | --------: | -------: | ------: |
|   FP16   |  6.339475 | 0.000002 |  99.941 |
|   Q8_0   |  6.345049 | 0.000854 |  98.503 |
|  E3M4_Q  |  6.345332 | 0.001324 |  98.045 |
|  E3M4  |  6.346416 | 0.001562 |  97.923 |
|  E4M3_Q  |  6.361256 | 0.004980 |  96.182 |
|   Q6_K   |  6.373544 | 0.006197 |  95.778 |
|  E4M3  |  6.374483 | 0.006227 |  95.598 |
|  Q5_K_M  |  6.397938 | 0.010158 |  94.834 |
|  Q5_K_S  |  6.437732 | 0.013509 |  94.231 |
|  E5M2   |  6.493300 | 0.022102 |  92.156 |
|  Q4_K_M  |  6.489127 | 0.026584 |  92.176 |

(Yes as you see NVIDIA/INTEL/AMD was wrong not add E3M4 in hardware.)

High speed FP8-CPU kernel is WIP but need this PR https://github.com/ggerganov/llama.cpp/pull/10446 for clean/simple 
 integration. 

Do you want we wait for this kernel before add FP8 ?